new file:   FRNN0D.h 	new file:   FRNN0D1.h 	new file:   FRNN0D_test_suite.c 	new file:   FRNN1D.h 	new file:   FRNN1D1.h 	new file:   FRNN1D_test_suite.c 	new file:   conf-0D.yaml 	new file:   conf-1D.yaml 	modified:   conf.yaml 	renamed:    sample_model.h5 -> example/model.h5 	deleted:    frnn_model2.h 	deleted:    frnn_model2_test_suite.c 	deleted:    frnn_model_test_suite.c 	renamed:    frnn_model.h -> frnn_stateful.h 	new file:   include/Makefile 	renamed:    include/k2c_activations.h -> include/k2c_activations.c 	renamed:    include/k2c_convolution_layers.h -> include/k2c_convolution_layers.c 	renamed:    include/k2c_core_layers.h -> include/k2c_core_layers.c 	renamed:    include/k2c_embedding_layers.h -> include/k2c_embedding_layers.c 	renamed:    include/k2c_helper_functions.h -> include/k2c_helper_functions.c 	modified:   include/k2c_include.h 	renamed:    include/k2c_merge_layers.h -> include/k2c_merge_layers.c 	renamed:    include/k2c_normalization_layers.h -> include/k2c_normalization_layers.c 	renamed:    include/k2c_pooling_layers.h -> include/k2c_pooling_layers.c 	renamed:    include/k2c_recurrent_layers.h -> include/k2c_recurrent_layers.c 	new file:   include/k2c_tensor_include.h 	new file:   include/libk2c.a 	modified:   keras2c/__main__.py 	modified:   keras2c/check_model.py 	modified:   keras2c/frnn_model_loader.py 	new file:   keras2c/keras2c/__init__.py 	new file:   keras2c/keras2c/__main__.py 	new file:   keras2c/keras2c/check_model.py 	new file:   keras2c/keras2c/frnn_model_loader.py 	new file:   keras2c/keras2c/io_parsing.py 	new file:   keras2c/keras2c/keras2c_main.py 	new file:   keras2c/keras2c/layer2c.py 	new file:   keras2c/keras2c/make_test_suite.py 	new file:   keras2c/keras2c/weights2c.py 	modified:   keras2c/keras2c_main.py 	modified:   keras2c/layer2c.py 	modified:   keras2c/make_test_suite.py 	modified:   keras2c/weights2c.py 	deleted:    load_frnn_model.py 	new file:   model.128801637189077864218421079770459754035._epoch_.24.h5 	new file:   model.209738965707153801274180429082834389454._epoch_.7.h5 	renamed:    conf.yaml -> old/conf-0d.yaml 	renamed:    model.293404425887829931850197837721263233440._epoch_.1.h5 -> old/model_epoch1.0d.h5 	modified:   scratch/scratch.c 	modified:   tests/test_core_layers.py 	modified:   tests/test_recurrent_layers.py 	new file:   website/README.rst

+++ tests/test_core_layers.py
@@ -24,12 +24,25 @@ def build_and_run(name):
         ccflags = '-g -Og -std=c99 -fprofile-arcs -ftest-coverage -I./include/'
     else:
         ccflags = '-Ofast -std=c99 -I./include/'
-    cc = 'gcc ' + ccflags + ' -o ' + name + ' ' + name + '_test_suite.c -lm'
-    subprocess.run(cc.split())
-    rcode = subprocess.run(['./' + name]).returncode
-    if not rcode and not os.environ.get('CI'):
-        subprocess.run('rm ' + name + '*', shell=True)
-    return rcode
+    cc = 'gcc ' + ccflags + '-I' + name + '.h ' + ' -o ' + name + ' ' + \
+        './include/k2c_activations.c ' + \
+        './include/k2c_convolution_layers.c ' + \
+        './include/k2c_core_layers.c ' + \
+        './include/k2c_embedding_layers.c ' + \
+        './include/k2c_helper_functions.c ' + \
+        './include/k2c_merge_layers.c ' + \
+        './include/k2c_normalization_layers.c ' + \
+        './include/k2c_pooling_layers.c ' + \
+        './include/k2c_recurrent_layers.c ' + \
+        name + '.c ' + name + '_test_suite.c -lm'
+    buildcode = subprocess.run(cc.split()).returncode
+    if not buildcode:
+        rcode = subprocess.run(['./' + name]).returncode
+        if not rcode and not os.environ.get('CI'):
+            subprocess.run('rm ' + name + '*', shell=True)
+        return rcode
+    else:
+        return buildcode
 
 
 class TestCoreLayers(unittest.TestCase):
@@ -194,6 +207,26 @@ class TestNormalization(unittest.TestCase):
                                             scale=False, center=False)(a)
         model = keras.models.Model(inputs=a, outputs=b)
         name = 'test___BatchNorm4' + str(int(time.time()))
+        keras2c_main.k2c(model, name)
+        rcode = build_and_run(name)
+        self.assertEqual(rcode, 0)
+
+
+class TestSharedLayers(unittest.TestCase):
+    """tests for shared layers"""
+
+    def test_SharedLayer1(self):
+        inshp = (10, 20)
+        xi = keras.layers.Input(inshp)
+        x = keras.layers.Dense(20, activation='relu')(xi)
+        yi = keras.layers.Input(inshp)
+        y = keras.layers.Dense(20, activation='relu')(yi)
+        f = keras.layers.Dense(30, activation='relu')
+        x = f(x)
+        y = f(y)
+        z = keras.layers.Add()([x, y])
+        model = keras.models.Model(inputs=[xi, yi], outputs=z)
+        name = 'test___SharedLayer1' + str(int(time.time()))
         keras2c_main.k2c(model, name)
         rcode = build_and_run(name)
         self.assertEqual(rcode, 0)
